Function dcs_decoder::decoder

pub fn decoder(consent_string: &str) -> HashMap<&str, DecodedType>

Decodes a Didomi Consent string into a map of key-value pairs.

Arguments

consent_string - The consent string to decode.

Returns

A HashMap where keys from Didomi Consent String Schema fields:

{ 
    "consent_string_type": "didomi_consent_string",
    "specification_version": 2,
    "fields": [
    {
      "type": "version",
      "description": "Didomi Consent String Specification Version, 6 bits number - 000010 or 'C' in Base64 for v2",
      "key": "version",
      "value": 2
    },
    {
      "type": "uuid",
      "description": "User ID (UUID)",
      "key": "user_id"
    },
    {
      "type": "date",
      "description": "UTC epoch time format when the Consent String was initially created",
      "key": "created"
    },
    {
      "type": "date",
      "description": "UTC epoch time format when the Consent String was last time updated",
      "key": "updated"
    },
    {
      "type": "u16",
      "description": "Regulation ID that was used to encode user status",
      "key": "regulation_id"
    },
    {
      "type": "date",
      "description": "Last sync date",
      "key": "sync",
      "optional": true
    },
    {
      "type": "enabled_disabled_ids",
      "description": "User Status for purposes opt-in",
      "key": "purposes_optin",
      "variants": [
        "bit_field_2_bits",
        "ranges_u16",
        "ranges_fibonacci"
      ]
    },
    {
      "type": "enabled_disabled_ids",
      "description": "User Status for purposes opt-out",
      "key": "purposes_optout",
      "variants": [
        "bit_field_2_bits",
        "ranges_u16",
        "ranges_fibonacci"
      ]
    },
    {
      "type": "enabled_disabled_ids",
      "description": "User Status for vendors opt-in",
      "key": "vendors_optin",
      "variants": [
        "bit_field_2_bits",
        "ranges_u16",
        "ranges_fibonacci"
      ]
    },
    {
      "type": "enabled_disabled_ids",
      "description": "User Status for vendors opt-out",
      "key": "vendors_optout",
      "variants": [
        "bit_field_2_bits",
        "ranges_u16",
        "ranges_fibonacci"
      ]
    }
  ]
}

Example

 
    let consent_string = "BGHWv4UYba5-dZnABdKu__D6iWHsD6iWHsJ9iee2BAAilKiABFKVQAACIscAAAiLGAAA";
    let user_status = decoder(consent_string);
    match user_status.get("purposes_optout") {
        Some(value) => {
            match value {
                DecodedType::RangesU16(user_status_item) |
                DecodedType::RangesFibonacci(user_status_item) |
                DecodedType::BitField2Bits(user_status_item) => {
                    assert_eq!(user_status_item.enabled, &[1, 2, 6, 7, 8]);
                    assert_eq!(user_status_item.disabled, &[3, 4])
                }
                _ => panic!("Failed to decode [purposes_optout] field with value")
            }
        }
        None => panic!("Mandatory [purposes_optout] is missing from the decoded result")
   }
   

See Also

DecodedType

UserStatusItem